Add filter_types_map
